Split routes into config and constants
