+++ src/app/routes/index.ts
@@ -1,0 +1,2 @@
+export { APP_ROUTES, APP_ROUTE_PARAMS } from './constants'
+export { routes } from './config'
